feat(context): Implement lookup through parent contexts
This commit makes the MustacheWriteVisitor context an actual stack. This has the effect that whenever a new context is introduced, it is pushed onto the context stack instead, allowing lookups on the context to traverse the ancestor contexts in reverse order, i.e. from parent to grandparent and so on.